+++ lib/glob.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require_relative "glob/map"
-require_relative "glob/query"
+require_relative "glob/object"
 require_relative "glob/matcher"
 require_relative "glob/symbolize_keys"
 require_relative "glob/version"
@@ -14,6 +14,6 @@ module Glob
   end
 
   def self.new(target = {})
-    Query.new(target)
+    Object.new(target)
   end
 end